requirements.txt
aiohttp==3.6.2
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.4
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
dash==1.4.1
dash-core-components==1.3.1
dash-html-components==1.0.1
dash-renderer==1.1.2
dash-table==4.4.1
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-jquery==3.1.0
django-location-field==2.1.0
django-tinymce==3.0.2
django-tinymce4-lite==1.8.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.4.0
future==0.18.2
glob2==0.7
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.16.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gspread==3.6.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsmin==2.2.2
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.1
kiwisolver==1.2.0
libarchive-c==2.9
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.6
nltk==3.5
notebook==6.0.3
numpy==1.18.1
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
plotly==4.7.1
polyline==1.4.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.12.2
psutil==5.7.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.13.0
PyQt5-sip==4.19.18
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==18.1.1
QtPy==1.9.0
rastermap==0.1.0
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
ruamel-yaml==0.15.87
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
slackclient==2.7.2
soupsieve==2.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.1.0
win-inet-pton==1.1.0
wincertstore==0.2
xmltodict==0.12.0
yarl==1.4.2
zipp==3.1.0
